fix compiler warnings

+++ include/vcpkg/commands.extract.h
@@ -12,7 +12,7 @@ namespace vcpkg::Commands
         std::vector<Path> proximate_to_temp;
     };
 
-    std::vector<std::pair<Path, Path>> strip_map(const ExtractedArchive& archive, int num_leading_dir);
+    std::vector<std::pair<Path, Path>> strip_map(const ExtractedArchive& archive, size_t num_leading_dir);
     size_t get_common_prefix_count(std::vector<Path> paths);
     void command_extract_and_exit(const VcpkgCmdArguments& args, const VcpkgPaths& paths);
 }
